refactor(race): Make public api for entity ui segment
- Made public api for entity ui segment
