+++ packages/client/src/entities/race/ui/index.ts
@@ -1,0 +1,1 @@
+export { RaceListCard } from './list-card';
